# swagger.yaml
openapi: 3.0.0
info:
  title: 'API Livros'
  description: 'Documentação da API Livros'
  version: 1.0.0
servers:
  -
    url: 'http://localhost/Api-Livros'
    description: 'Servidor local'
paths:
  /register:
    post:
      tags:
        - Auth
      summary: 'Realiza o cadastro de um novo usuário'
      operationId: 6ca1953c104d60936b62117784734f90
      requestBody:
        required: true
        content:
          application/json:
            schema:
              required:
                - nome
                - email
                - senha
              properties:
                nome:
                  type: string
                  example: 'Luiz Felipe'
                  maxLength: 255
                email:
                  type: string
                  format: email
                  example: teste@email.com
                senha:
                  type: string
                  format: password
                  example: 12345678Ab@
              type: object
      responses:
        '201':
          description: 'Usuário registrado com sucesso'
          content:
            application/json:
              schema:
                properties:
                  mensagem: { type: string, example: 'Usuário registrado com sucesso' }
                type: object
        '400':
          description: 'Dados inválidos ou email já cadastrado'
          content:
            application/json:
              schema:
                properties:
                  mensagem: { type: string, example: 'Email já cadastrado' }
                type: object
        '500':
          description: 'Erro interno ao cadastrar usuário'
          content:
            application/json:
              schema:
                properties:
                  mensagem: { type: string, example: 'Erro ao cadastrar usuário' }
                type: object
  /login:
    post:
      tags:
        - Auth
      summary: 'Realiza login do usuário'
      operationId: f6b87925c502a2f52803eb2945e0b2f6
      requestBody:
        required: true
        content:
          application/json:
            schema:
              required:
                - email
                - senha
              properties:
                email:
                  type: string
                  format: email
                  example: teste@email.com
                senha:
                  type: string
                  format: password
                  example: 12345678Ab@
              type: object
      responses:
        '200':
          description: 'Login realizado com sucesso'
          content:
            application/json:
              schema:
                properties:
                  mensagem: { type: string, example: 'Login realizado com sucesso' }
                  access_token: { type: string, example: jwt_aqui }
                  UUID: { type: string, example: e04230085f34fcdc518137ac826725 }
                  nome: { type: string, example: 'Luiz Felipe' }
                  email: { type: string, format: email, example: teste@email.com }
                  foto_perfil: { type: string, example: null, nullable: true }
                type: object
        '400':
          description: 'Erro de validação na requisição'
          content:
            application/json:
              schema:
                properties:
                  mensagem: { type: string, example: 'Dados inválidos' }
                type: object
        '401':
          description: 'Credenciais inválidas'
          content:
            application/json:
              schema:
                properties:
                  mensagem: { type: string, example: 'Email ou senha inválidos' }
                type: object
  /perfil:
    get:
      tags:
        - Auth
      summary: 'Retorna os dados do perfil do usuário autenticado'
      operationId: 11b2b4bdfe5b35bed6f42bed010cae98
      responses:
        '200':
          description: 'Perfil acessado com sucesso'
          content:
            application/json:
              schema:
                properties:
                  success: { type: boolean, example: true }
                  mensagem: { type: string, example: 'Perfil acessado com sucesso' }
                  id_usuario: { type: integer, example: 1 }
                  usuario: { properties: { id_usuario: { type: integer, example: 1 }, UUID: { type: string, example: e04230085f34fcdc518137ac826725 }, nome: { type: string, example: 'Luiz Felipe' }, email: { type: string, format: email, example: teste@email.com } }, type: object }
                type: object
        '401':
          description: 'Token ausente, inválido ou sem permissão'
          content:
            application/json:
              schema:
                properties:
                  mensagem: { type: string, example: 'Token inválido para acesso' }
                type: object
      security:
        -
          bearerAuth: []
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      bearerFormat: JWT
      scheme: bearer
...